# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "N100 Financial Intelligence Platform",
    "description": "FastAPI Backend for N100 Financial Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/peers/{group_name}": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peer Group Companies and Percentiles",
        "description": "Retrieve percentile rankings for all 10 core peer metrics across companies in a peer group.",
        "operationId": "get_peer_group_details_api_v1_peers__group_name__get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Peer group percentile data returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerGroupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Peer group not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peer Radar Comparison Data",
        "description": "Retrieve 8 radar axis metrics for requested company, peer group average, and benchmark company.",
        "operationId": "get_company_peer_compare_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Radar comparison data returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadarCompareResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company ticker not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portfolio/stats": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Portfolio KPI Percentile Statistics",
        "description": "Retrieve statistical percentiles (P10, P25, P50, P75, P90, Mean, Std) for core financial KPIs across the company universe.",
        "operationId": "get_portfolio_stats_api_v1_portfolio_stats_get",
        "responses": {
          "200": {
            "description": "Portfolio statistics returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioStatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KPIPercentileStatItem": {
        "properties": {
          "kpi": {
            "type": "string",
            "title": "Kpi",
            "description": "Financial KPI identifier"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count",
            "description": "Number of valid company samples"
          },
          "P10": {
            "type": "number",
            "title": "P10",
            "description": "10th Percentile value"
          },
          "P25": {
            "type": "number",
            "title": "P25",
            "description": "25th Percentile value"
          },
          "P50": {
            "type": "number",
            "title": "P50",
            "description": "50th Percentile (Median) value"
          },
          "P75": {
            "type": "number",
            "title": "P75",
            "description": "75th Percentile value"
          },
          "P90": {
            "type": "number",
            "title": "P90",
            "description": "90th Percentile value"
          },
          "Mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean",
            "description": "Mean average value"
          },
          "Std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Std",
            "description": "Standard deviation"
          }
        },
        "type": "object",
        "required": [
          "kpi",
          "P10",
          "P25",
          "P50",
          "P75",
          "P90"
        ],
        "title": "KPIPercentileStatItem"
      },
      "PeerCompanyPercentileItem": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Company name"
          },
          "percentiles": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Percentiles",
            "description": "Dictionary mapping metric to percentile rank"
          },
          "metric_details": {
            "items": {
              "$ref": "#/components/schemas/PeerPercentileMetricItem"
            },
            "type": "array",
            "title": "Metric Details",
            "description": "Detailed metric list with values and percentiles"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "company_name",
          "percentiles",
          "metric_details"
        ],
        "title": "PeerCompanyPercentileItem"
      },
      "PeerGroupResponse": {
        "properties": {
          "peer_group_name": {
            "type": "string",
            "title": "Peer Group Name",
            "description": "Peer group identifier"
          },
          "company_count": {
            "type": "integer",
            "title": "Company Count",
            "description": "Total companies in peer group"
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/PeerCompanyPercentileItem"
            },
            "type": "array",
            "title": "Companies",
            "description": "List of companies with 10 metric percentiles"
          }
        },
        "type": "object",
        "required": [
          "peer_group_name",
          "company_count",
          "companies"
        ],
        "title": "PeerGroupResponse"
      },
      "PeerPercentileMetricItem": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric",
            "description": "Financial metric name"
          },
          "metric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Value",
            "description": "Actual metric value"
          },
          "percentile_rank": {
            "type": "number",
            "title": "Percentile Rank",
            "description": "Percentile rank within peer group (0.0 to 1.0)"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "percentile_rank"
        ],
        "title": "PeerPercentileMetricItem"
      },
      "PortfolioStatsResponse": {
        "properties": {
          "total_kpis": {
            "type": "integer",
            "title": "Total Kpis",
            "description": "Total core KPIs analyzed"
          },
          "stats": {
            "items": {
              "$ref": "#/components/schemas/KPIPercentileStatItem"
            },
            "type": "array",
            "title": "Stats",
            "description": "Percentile statistics list for core KPIs"
          }
        },
        "type": "object",
        "required": [
          "total_kpis",
          "stats"
        ],
        "title": "PortfolioStatsResponse"
      },
...
      "RadarCompareResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "Requested company ticker"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Requested company name"
          },
          "peer_group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Group Name",
            "description": "Assigned peer group"
          },
          "benchmark_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Ticker",
            "description": "Benchmark company ticker for this peer group"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metrics",
            "description": "List of 8 radar axis metrics"
          },
          "company_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Company Values",
            "description": "8 metric values for requested company"
          },
          "peer_average": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Peer Average",
            "description": "8 metric average values across peer group"
          },
          "benchmark_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Benchmark Values",
            "description": "8 metric values for benchmark company"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name",
          "metrics",
          "company_values",
          "peer_average",
          "benchmark_values"
        ],
        "title": "RadarCompareResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}